fix: resolve final 3 sign-compare errors in test_output
- test_midi_beat_clock.cpp: lines 330, 337 (clocks_sent comparisons)
- test_relay_output.cpp: line 105 (pulse_count comparison)
- test_network_transport.cpp: line 337 (packets_sent with cast)

+++ test/test_output/test_network_transport.cpp
@@ -334,5 +334,5 @@ TEST_F(NetworkTransportTest, BufferOverflow_ProtectedAgainstBurst) {
     
     // And: Stats are consistent
     auto stats = controller->getNetworkStats();
-    EXPECT_EQ(stats.packets_sent, success_count);
+    EXPECT_EQ(stats.packets_sent, static_cast<uint32_t>(success_count));
 }
